backend/functions/requirements.txt
annotated-types==0.6.0
anyio==4.1.0
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudevents==1.9.0
cmake==3.26.4
contourpy==1.1.0
cryptography==41.0.7
cvxpy==1.3.2
cycler==0.11.0
deprecation==2.1.0
distro==1.8.0
docopt==0.6.2
ecos==2.0.12
firebase-admin==6.3.0
firebase-functions==0.2.0
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.41.0
functions-framework==3.5.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-events==0.11.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
msgpack==1.0.7
numpy==1.25.1
oauthlib==3.2.2
openai==1.3.9
osqp==0.6.3
packaging==23.1
Pillow==10.0.0
pipreqs==0.4.13
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0.1
qdldl==0.1.7.post0
rapidfuzz==3.6.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.1
scs==3.2.3
six==1.16.0
sniffio==1.3.0
thefuzz==0.20.0
tqdm==4.66.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.0.4
watchdog==3.0.0
Werkzeug==3.0.1
yarg==0.1.9
